update comment cud methods. working now
