+++ src/main/java/com/revature/domain/QuestionEval.java
@@ -140,7 +140,7 @@ public class QuestionEval implements Serializable {
 	
 	@Override
 	public String toString() {
-		return "QuestionEval{" + "id=" + id + ", communicationScore=" + communicationScore + ", knowledgeScore=" + knowledgeScore + ", eval=" + eval + ", questionPool=" + questionPool + ", comments=" + comments + '}';
+		return "QuestionEval{" + "id=" + id + ", communicationScore=" + communicationScore + ", knowledgeScore=" + knowledgeScore + ", eval.id=" + eval.getId() + ", questionPool=" + questionPool + ", comments=" + comments + '}';
 	}	
 	
 }
